% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{column_pairs}
\alias{column_pairs}
\title{Compute pairs of colnames based on the following architecture:
Suppose a dataframe with column ordered as follow : c("A", "B", "C", "D")
The pairs will be computed as follow
AA AB AC
BA BB BC
CA CB CC}
\usage{
column_pairs(data)
}
\arguments{
\item{data}{Dataframe}
}
\value{
A list containing each pair. Would look like this
[\link{1}]
\link{1} "A" "A"

[\link{2}]
\link{1} "A" "B"

[\link{3}]
\link{1} "A" "C"

[\link{4}]
\link{1} "B" "A"

[\link{5}]
\link{1} "B" "B"

[\link{6}]
\link{1} "B" "C"

[\link{7}]
\link{1} "C" "A"

[\link{8}]
\link{1} "C" "B"
[\link{9}]
\link{1} "C" "C"
}
\description{
Compute pairs of colnames based on the following architecture:
Suppose a dataframe with column ordered as follow : c("A", "B", "C", "D")
The pairs will be computed as follow
AA AB AC
BA BB BC
CA CB CC
}
