% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{configure_theme}
\alias{configure_theme}
\title{Set theme programatically conditional on arguments.}
\usage{
configure_theme(repeat_labels, repeat_text, contigous_l, contigous_b)
}
\arguments{
\item{repeat_text}{See \code{\link[=pairgrid]{pairgrid()}}}

\item{contigous_l}{A boolean indicating whether subplot is contigous to the left side.}

\item{contigous_b}{A boolean indicating whether subplot is contigous to the bottom.}

\item{repeat_label}{See \code{\link[=pairgrid]{pairgrid()}}}
}
\description{
Set theme programatically conditional on arguments.
}
