% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{global_ylim}
\alias{global_ylim}
\title{Description analog to \code{\link[=global_xlim]{global_xlim()}} but for y axis.}
\usage{
global_ylim(data, y)
}
\description{
Description analog to \code{\link[=global_xlim]{global_xlim()}} but for y axis.
}
