% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{mapplot}
\alias{mapplot}
\title{In order to to construct a full pair matrix. This function has to be called
for each part (lower corner, upper corner, diagonal). The call on diagonal should be last
in order to have access to neighbour subplots metadata.}
\usage{
mapplot(
  data,
  mapping,
  pairs,
  grobs,
  indices,
  func,
  no_upper,
  no_lower,
  xlim_func = NULL,
  ylim_func = NULL,
  diag_share_ylim = TRUE,
  common_scale = NULL,
  check.overlap = TRUE,
  neighbours = NULL,
  ...
)
}
\arguments{
\item{xlim_func}{See \code{\link[=pairgrid]{pairgrid()}} parameter \code{common_xlim} (equivalent).}

\item{ylim_func}{A function to set y axis limits. If set in conjunction with \code{diag_share_ylim} set to TRUE,
plots on the ylimit, will have their ylim set in common.}

\item{diag_share_ylim}{If set to TRUE. Y axis limits will also use \code{ylim_func}, except for diagonal only pairplot. In this case,
will be normalized to unit length if the provided function implements unit scale normalization. See function
\link{pair_geom_histogram} for an example.}

\item{check.overlap}{See\code{\link[ggplot2:guide_axis]{ggplot2::guide_axis()}}}

\item{neighbours}{A map (named list) or \code{NULL}.
Map of the form: \verb{subplot_index=list(t=<int|NULL>, b=<int|NULL>, l=<int|NULL>, r=<int|NULL>)}
If not NULL, the function call
\code{func()} will include the data in argument \code{neighbour_ggdata} from the neighbour(s): \code{func(neighbour_ggdata=list(t=ggplot_build(grobs_index[[t_index]])), etc.)}.
There are no checks whether the selected grobs exist. Checks are reserved to function calling \code{mapplot}.}
}
\description{
In order to to construct a full pair matrix. This function has to be called
for each part (lower corner, upper corner, diagonal). The call on diagonal should be last
in order to have access to neighbour subplots metadata.
}
