% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{pair_geom_histogram}
\alias{pair_geom_histogram}
\title{Min max histogram}
\usage{
pair_geom_histogram(
  data,
  mapping,
  neighbour_ggdata = NULL,
  unit_y = FALSE,
  stat = "bin",
  bins = 10,
  binwidth = function(x) {
     bin_width_auto(x, na.rm = TRUE)
 },
  ...
)
}
\arguments{
\item{unit_y}{Whether to scale y (min-max) to interval \link{0, 1}.}

\item{min_max}{A function for transforming the data.
Those values are used as boundaries. If only x or y is provided. Scaling is done on one dimension only.
for min max scaling.}
}
\description{
Min max histogram
}
