% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairplot.R
\name{pair_geom_histogram}
\alias{pair_geom_histogram}
\title{Min max histogram}
\usage{
pair_geom_histogram(
  data,
  mapping,
  unit_y = FALSE,
  stat = "bin",
  bins = 10,
  binwidth = function(x) {
     bin_width_auto(x, na.rm = TRUE)
 },
  ...
)
}
\arguments{
\item{unit_y}{Whether to scale y (min-max) to interval $\link{0, 1}$. If FALSE, min-max occurs but interval is set to $\link{min(y), max(y)}$,}
}
\description{
Min max histogram
}
